tests: subsys: cmsis_rtos_v2: revise the timeout in mutex and sem tasks
In test_mutex_lock_timeout and test_semaphore tests, the threads are
created. In the threads, the resource is occupied for three times. one is
wait for forever, another is set 10 ticks timeout and the other is 5 ticks
timeout. However, the 5 and 10 ticks is not enough. This may block the
following works and cause system halt. This change revises timeout to 20
and 10 ticks to fix this issue.

Signed-off-by: Ren Chen <Ren.Chen@ite.com.tw>