drivers: usb: usb_dc_it82xx2: correct the extend endpoint control
There are some issues with the extended endpoint settings. The incorrect
setting leads to the chip being unable to respond with NAK when the host
polls the extended endpoint for data transfers. Additionally, the controls
for the extended endpoint's ISO and PID data sequence are also incorrect.
This commit addresses these issues to properly support extended endpoint
transactions.

Change-Id: Ib7dc26a0904d8c4babf991e5e685403bcf8e1285
Signed-off-by: Ren Chen <Ren.Chen@ite.com.tw>