apa102: fix end frame
end frame is used to supply clock pulses so that data goes to last
LED in the chain. Thus, it depends on the number of LEDs in the chain.
Previously, the number of ones sent into the end frame was
hard-coded and limited the usage of the driver to 64 LEDs in the
strip.

(cherry picked from commit 89594e3ada7756dc1dc790d8bd1666d86b08b3eb)

Original-Signed-off-by: Cyril Fougeray <cyril.fougeray@toolsforhumanity.com>
GitOrigin-RevId: 89594e3ada7756dc1dc790d8bd1666d86b08b3eb
Cr-Build-Id: 8725724783141066929
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8725724783141066929
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ie8566437f56e8e287fead5b819183c5506357c69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6176148
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>