doc: _extensions: fix :zephyr:board: role with Intersphinx
Resolving the target for a zephyr:board role when using Interpshinx
needs to account for the fact that a board object might be in the
domain (it's part of the catalog), and yet not have any actual docname
associated with it for there's no page in the "local" docs
that defines a ..zephyr:board:: documentation for it.
Handle that case so that ZephyrDomain#resolve_xref yields to Intersphinx
that will eventually find the target in the inventory for
docs.zephyrproject.org.

Fixes https://github.com/zephyrproject-rtos/zephyr/issues/83737

(cherry picked from commit 92271d173fe23f464bf8ce0fe2dc7ce9258fd9dc)

Original-Signed-off-by: Benjamin Cabé <benjamin@zephyrproject.org>
GitOrigin-RevId: 92271d173fe23f464bf8ce0fe2dc7ce9258fd9dc
Cr-Build-Id: 8725543589450806177
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8725543589450806177
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0c11837239dd6a9949751b2627f2878d3b030cf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6181114
Commit-Queue: Jonathon Murphy <jpmurphy@google.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>