doc: gh_utils: ignore added-but-not-committed files
If the file is added to the index, but not committed yet, the 'git'
command will return an empty string. This patch checks for this case and
treats it as if the file was not tracked at all.

(cherry picked from commit f3bf6357c799191c230142ecb04994c0f232ed16)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: f3bf6357c799191c230142ecb04994c0f232ed16
Cr-Build-Id: 8725543589450806177
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8725543589450806177
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I47f21682f9a5543c2b0f00417122d82953e9a43c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6181165
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Commit-Queue: Jonathon Murphy <jpmurphy@google.com>
Tested-by: Keith Short <keithshort@chromium.org>